Gullintanni v0.1.0 Gullintanni.Config

Utility module for dealing with application configuration settings.

Types

t()

t() :: Keyword.t

The configuration settings type

Functions

load(key)

Returns the value of the application configuration specified by key.

Returns [] if no configuration was found.

load(key, subkey)

load(atom, t) :: t

load(atom, atom) :: t

Returns the value of the subkey of the parent application configuration specified by key.

Returns [] if no configuration was found.

parse_runtime_settings(config)

parse_runtime_settings(t) :: t

Transforms all the {:system, "ENV_VAR"} tuples into their respective values grabbed from the process environment.

settings_present?(keys, config)

settings_present?([atom], t) :: boolean

Returns true if all of the keys are present in the config, otherwise false.

Logs an error for any missing settings.