Gullintanni Bot v0.1.0 Gullintanni.Listener.Socket

Utility module for handling TCP socket parsing and validation.

Types

address :: String.t | charlist | :inet.ip_address

The IP address of a socket

port_number :: 0..65535

The port number of a socket

t :: %Gullintanni.Listener.Socket{address: address, port: port_number}

The socket type

Functions

new(address, port)

Specs

new(address, port_number) ::
  {:ok, t} |
  {:error, :invalid_address} |
  {:error, :invalid_port}

Creates a new socket with the given address and port.

Returns {:ok, socket} if the address and port are valid, returns {:error, reason} otherwise. A valid address is anything that can be parsed by :inet.parse_address/1, and a valid port must be an integer in the range of 0..65535.

Examples

iex> {:ok, socket} = Gullintanni.Listener.Socket.new("0.0.0.0", 80)
iex> socket
#Socket<0.0.0.0:80>

iex> {:ok, socket} = Gullintanni.Listener.Socket.new("fe80::204:acff:fe17:bf38", 80)
iex> socket
#Socket<[FE80::204:ACFF:FE17:BF38]:80>

iex> Gullintanni.Listener.Socket.new("100.200.300.400", 80)
{:error, :invalid_address}

iex> Gullintanni.Listener.Socket.new("0.0.0.0", 99999)
{:error, :invalid_port}