Gullintanni v0.1.0 Gullintanni.Pipeline

Defines a Gullintanni build pipeline.

A pipeline is uniquely identified by a provider plus a repository, and encapsulates the state necessary to coordinate builds and merges.

Types

t :: %Gullintanni.Pipeline{config: Gullintanni.Config.t, provider: Gullintanni.Provider.t, queue: Gullintanni.Queue.t, repo: Gullintanni.Repo.t, worker: Gullintanni.Worker.t}

The pipeline type

Functions

download_merge_requests(pipeline)

Specs

download_merge_requests(t) :: [Gullintanni.MergeRequest.t]

Downloads a list of the repository’s open merge requests.

id(pipeline)

Specs

id(t) :: String.t

Returns a string representation that uniquely identifies a pipeline.

init_queue(pipeline)

Specs

init_queue(t) :: t

Initializes the pipeline’s queue by downloading a list of the repository’s open merge requests.

This replaces any existing queue that may be stored in the pipeline.

load(name, opts \\ [])

Specs

load(atom, Keyword.t) :: {:ok, t} | :error

Creates a pipeline with settings loaded from the application configuration.

The settings are loaded from the named pipeline’s config, as set in the application configuration. Any options specified in opts will then override those settings.

Returns {:ok, pipeline} if the configuration is valid, otherwise :error.

valid_config?(config)

Specs

valid_config?(Gullintanni.Config.t) :: boolean

Returns true if all required pipeline configuration values exist in config, otherwise false.

whoami(pipeline)

Specs

whoami(t) :: String.t

Returns the provider account’s effective user identity.