Gullintanni v0.1.0 Gullintanni.Pipeline.Supervisor

Supervises the Gullintanni build pipline processes.

Functions

init()

Callback invoked to start the supervisor and during hot code upgrades.

Options

:strategy - the restart strategy option. Only :one_for_one is supported by dynamic supervisors.

:max_restarts - the maximum amount of restarts allowed in a time frame. Defaults to 3 times.

:max_seconds - the time frame in which :max_restarts applies in seconds. Defaults to 5 seconds.

:max_dynamic - the maximum number of children started under the supervisor via start_child/2. Defaults to infinity children.

:subscribe_to - a list of producers to subscribe to. Each element represents the producer or a tuple with the producer and the subscription options. e.g. [Producer] or [{Producer, max_demand: 10, min_demand: 20}]

Callback implementation for Experimental.DynamicSupervisor.init/1.

start_link()

start_pipeline(config)