Gullintanni Bot v0.1.0 Gullintanni.Provider behaviour

This module specifies the API that a provider is required to implement.

Providers are adapter modules that interface with a Git hosting service.

Types

config :: Keyword.t

t :: module

Callbacks

get_merge_requests(arg0, config)

Specs

get_merge_requests(Gullintanni.Repo.t, config) :: [Gullintanni.MergeRequest.t]

Downloads a list of the repository’s open merge requests.

parse_merge_request(map)

Specs

parse_merge_request(map) :: Gullintanni.MergeRequest.t

Converts raw upstream data into a merge request.

validate_config(config)

Specs

validate_config(config) :: :ok | no_return

Returns :ok if all required provider configuration values exist in config, otherwise raises an ArgumentError exception.

whoami(config)

Specs

whoami(config) :: String.t

Returns the account’s effective user identity.