Gullintanni v0.1.0 Gullintanni.Providers.GitHub

Provider adapter module for GitHub — https://github.com/.

Functions

download_merge_requests(repo, config)

Downloads a list of the repository’s open merge requests.

Callback implementation for Gullintanni.Provider.download_merge_requests/2.

parse_comment(data)

parse_comment(map) :: Gullintanni.Comment.t

parse_merge_request(data)

parse_merge_request(map) :: Gullintanni.MergeRequest.t

Converts raw upstream data into a merge request.

Callback implementation for Gullintanni.Provider.parse_merge_request/1.

parse_repo(data)

parse_repo(map) :: Gullintanni.Repo.t

post_comment(repo, thread, comment, config)

post_comment(Gullintanni.Repo.t, Gullintanni.MergeRequest.id, String.t, Gullintanni.Config.t) ::
  :ok |
  :error

Publishes a comment in response to the given message.

valid_config?(config)

Returns true if all required provider configuration values exist in config, otherwise false.

Callback implementation for Gullintanni.Provider.valid_config?/1.

whoami(config)

Returns the provider account’s effective user identity.

Callback implementation for Gullintanni.Provider.whoami/1.