Gullintanni v0.1.0 Gullintanni.Queue

Provides priority queues in a relatively efficient manner.

Queues are double-ended. The mental picture of a queue is a line of people (items) waiting for their turn. The queue front is the end with the item that has waited the longest. The queue rear is the end an item enters when it starts to wait. If instead using the mental picture of a list, the front is called head and the rear is called tail. A priority queue adds the idea that an item can skip ahead in line according to its importance (priority).

Each item in a priority queue is represented as a tuple with both its value and priority. Priorities must be integers, with higher integers being treated as higher in priority. Multiple items may exist with the same priority and are used in First-In First-Out (FIFO) order.

The data representing a priority queue as used by this module is to be regarded as opaque by other modules.

Types

data

The underlying data structure of a priority queue

item :: {value, priority}

A priority queue item

priority :: integer

The priority of an item

t :: %Gullintanni.Queue{items: data, size: non_neg_integer}

The priority queue type

value :: any

The value of an item

Functions

delete(queue, item)

Specs

delete(t, item) :: t

Deletes an item from the queue.

Returns a new queue which is a copy of queue but without item. If the item did not exist previously, returns the queue unmodified. If the item’s value occurs more than once at the specified priority, just the item with the highest relative priority is removed.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.delete(queue, {:charlie, 0})
#Queue<size: 2, front: {:alice, 0}, rear: {:bob, 0}>

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.delete(queue, {:dave, 0})
#Queue<size: 3, front: {:alice, 0}, rear: {:charlie, 0}>

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie, :alice])
iex> Gullintanni.Queue.delete(queue, {:alice, 0})
#Queue<size: 3, front: {:bob, 0}, rear: {:alice, 0}>

empty?(queue)

Specs

empty?(t) :: boolean

Returns true if the queue is empty, otherwise false.

Examples

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.empty?(queue)
true

iex> queue = Gullintanni.Queue.new([:alice])
iex> Gullintanni.Queue.empty?(queue)
false

insert(queue, item)

Specs

insert(t, item | value) :: t

Inserts an item into the queue.

The item can be provided as a raw value (which will be inserted using the default priority of 0), or in the {value, priority} format.

Returns a new queue which is a copy of queue but with the item’s value inserted at the specified priority. If there are existing values at the specified priority, the new value will have the lowest relative priority of those values.

Examples

iex> queue = Gullintanni.Queue.new([:alice])
iex> Gullintanni.Queue.insert(queue, :bob)
#Queue<size: 2, front: {:alice, 0}, rear: {:bob, 0}>

iex> queue = Gullintanni.Queue.new([:alice])
iex> Gullintanni.Queue.insert(queue, {:bob, 5})
#Queue<size: 2, front: {:bob, 5}, rear: {:alice, 0}>

move(queue, item, new_priority)

Specs

move(t, item, priority) :: t

Moves an item to a different priority in the queue.

Returns a new queue which is a copy of queue but with the item’s value moved to the specified priority. If the item did not exist previously, returns the queue unmodified. If the item’s value occurs more than once at the specified priority, just the item with the highest relative priority is moved.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.move(queue, {:charlie, 0}, 5)
#Queue<size: 3, front: {:charlie, 5}, rear: {:bob, 0}>

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.move(queue, {:dave, 0}, 5)
#Queue<size: 3, front: {:alice, 0}, rear: {:charlie, 0}>

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie, :alice])
iex> Gullintanni.Queue.move(queue, {:alice, 0}, 5)
#Queue<size: 4, front: {:alice, 5}, rear: {:alice, 0}>

new()

Specs

new :: t

Returns an empty priority queue.

Examples

iex> Gullintanni.Queue.new
#Queue<empty>

new(items)

Specs

new(Enum.t) :: t

Creates a new priority queue from a list of enumerable items.

Each item can be provided as a raw value (which will be inserted using the default priority of 0), or in the {value, priority} format.

Examples

iex> Gullintanni.Queue.new([:alice, :bob, :charlie])
#Queue<size: 3, front: {:alice, 0}, rear: {:charlie, 0}>

iex> Gullintanni.Queue.new([:alice, {:bob, 5}, :charlie])
#Queue<size: 3, front: {:bob, 5}, rear: {:charlie, 0}>

peek(queue)

Specs

peek(t) :: item | :empty

Returns the highest priority item in the queue.

Returns :empty if the queue is empty.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.peek(queue)
{:alice, 0}

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.peek(queue)
:empty

peek_r(queue)

Specs

peek_r(t) :: item | :empty

Returns the lowest priority item in the queue.

Returns :empty if the queue is empty.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.peek_r(queue)
{:charlie, 0}

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.peek_r(queue)
:empty

pop(queue)

Specs

pop(t) :: {t, item} | :empty

Returns and removes the highest priority item in the queue.

Returns :empty if the queue is empty.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> {new_queue, item} = Gullintanni.Queue.pop(queue)
iex> new_queue
#Queue<size: 2, front: {:bob, 0}, rear: {:charlie, 0}>
iex> item
{:alice, 0}

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.pop(queue)
:empty

pop_r(queue)

Specs

pop_r(t) :: {t, item} | :empty

Returns and removes the lowest priority item in the queue.

Returns :empty if the queue is empty.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> {new_queue, item} = Gullintanni.Queue.pop_r(queue)
iex> new_queue
#Queue<size: 2, front: {:alice, 0}, rear: {:bob, 0}>
iex> item
{:charlie, 0}

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.pop_r(queue)
:empty

size(queue)

Specs

size(t) :: non_neg_integer

Returns the number of items in the queue.

Examples

iex> queue = Gullintanni.Queue.new([:alice, :bob, :charlie])
iex> Gullintanni.Queue.size(queue)
3

iex> queue = Gullintanni.Queue.new
iex> Gullintanni.Queue.size(queue)
0