Gullintanni v0.1.0 Gullintanni.Worker behaviour

Specifies the API that a worker module is required to implement.

Workers are adapter modules that interface with a CI service.

Types

config :: Keyword.t

t :: module

Callbacks

valid_config?(config)

Specs

valid_config?(config) :: boolean

Returns true if all required worker configuration values exist in config, otherwise false.