Gullintanni v0.1.0 HardHat

A simple wrapper for the Travis CI API.

Types

response()

response() :: map | {status_code, binary}

The response to an HTTP request.

status_code()

status_code() :: integer

The HTTP status code.

Functions

get(client, path)

get(HardHat.Client.t, String.t) :: response

Issues an authenticated GET request to the given Travis CI API path.

Returns the decoded response body if the request was successful, otherwise {status_code, body}. Raises an exception in case of failure.

post(client, path, body \\ "")

post(HardHat.Client.t, String.t, HTTPoison.body) :: response

Issues an authenticated POST request to the given Travis CI API path.

Returns the decoded response body if the request was successful, otherwise {status_code, body}. Raises an exception in case of failure.

put(client, path, body \\ "")

put(HardHat.Client.t, String.t, HTTPoison.body) :: response

Issues an authenticated PUT request to the given Travis CI API path.

Returns the decoded response body if the request was successful, otherwise {status_code, body}. Raises an exception in case of failure.