Gullintanni v0.1.0 HardHat.Client

Defines a client for connecting to the Travis CI API.

A client is the combination of authentication credentials and an endpoint URL.

Types

auth()

auth() :: %{access_token: String.t}

The API authentication credentials

endpoint()

endpoint() :: String.t

The API endpoint URL

t()

t() :: %HardHat.Client{auth: auth, endpoint: endpoint}

The client type

Functions

new(auth, endpoint \\ "https://api.travis-ci.org/")

new(auth, endpoint) :: t

Creates a new client with the given auth and endpoint.