Gullintanni v0.1.0 Socket

Defines an Internet socket address.

A socket address is the combination of an IP address and a port number. This can be used along with a transport protocol in order to define an endpoint of a connection across a network.

Types

ip_address()

ip_address() :: String.t | charlist | :inet.ip_address

The IP address of a socket

port_number()

port_number() :: 0..65535

The port number of a socket

t()

t() :: %Socket{ip: :inet.ip_address, port: port_number}

The socket address type

Functions

new(ip, port)

new(ip_address, port_number) ::
  {:ok, t} |
  {:error, :invalid_ip_address | :invalid_port}

Creates a new socket address with the given ip and port.

Returns {:ok, socket} if the IP address and port number are valid, returns {:error, reason} otherwise. A valid IP address is anything that can be parsed by :inet.parse_address/1, and a valid port must be an integer in the range of 0..65535.

Examples

iex> {:ok, socket} = Socket.new("127.0.0.1", 80)
iex> socket
#Socket<127.0.0.1:80>

iex> {:ok, socket} = Socket.new("fe80::204:acff:fe17:bf38", 80)
iex> socket
#Socket<[FE80::204:ACFF:FE17:BF38]:80>
iex> socket.ip
{65152, 0, 0, 0, 516, 44287, 65047, 48952}
iex> socket.port
80

iex> Socket.new("100.200.300.400", 80)
{:error, :invalid_ip_address}

iex> Socket.new("0.0.0.0", 99999)
{:error, :invalid_port}