Gullintanni Bot v0.1.0 Gullintanni.Pipeline

Types

config :: Keyword.t

The raw pipeline configuration settings

name :: atom

A pipeline configuration name

t :: %Gullintanni.Pipeline{config: config, provider: Gullintanni.Provider.t, queue: Gullintanni.Queue.t, repo: Gullintanni.Repo.t}

The pipeline type

Functions

load(name, opts \\ [])

Specs

load(name, config) :: t

Creates a pipeline with settings loaded from the application configuration.

The settings are loaded from the named pipeline’s config, as set in the application configuration. Any options specified in opts will then override those settings.

validate_config(config)

Specs

validate_config(config) :: :ok | no_return

Returns :ok if all required pipeline configuration values exist in config, otherwise raises an ArgumentError exception.

whoami(pipeline)

Specs

whoami(t) :: String.t

Returns the provider’s effective user identity.