Gullintanni v0.1.0 Gullintanni.Provider behaviour

Specifies the API that a provider module is required to implement.

Providers are adapter modules that interface with a Git hosting service.

Types

t()

t() :: module

Callbacks

__domain__()

__domain__() :: String.t

Returns the domain name of the provider.

download_merge_requests(arg0, arg1)

download_merge_requests(Gullintanni.Repo.t, Gullintanni.Config.t) :: [Gullintanni.MergeRequest.t]

Downloads a list of the repository’s open merge requests.

parse_merge_request(map)

parse_merge_request(map) :: Gullintanni.MergeRequest.t

Converts raw upstream data into a merge request.

valid_config?(arg0)

valid_config?(Gullintanni.Config.t) :: boolean

Returns true if all required provider configuration values exist in config, otherwise false.

whoami(arg0)

whoami(Gullintanni.Config.t) :: String.t

Returns the provider account’s effective user identity.