Gullintanni v0.1.0 Gullintanni.Worker behaviour

Specifies the API that a worker module is required to implement.

Workers are adapter modules that interface with a CI service.

Types

t()

t() :: module

Callbacks

valid_config?(arg0)

valid_config?(Gullintanni.Config.t) :: boolean

Returns true if all required worker configuration values exist in config, otherwise false.

whoami(arg0)

whoami(Gullintanni.Config.t) :: String.t

Returns the worker account’s effective user identity.